refactor: moved the field for inclusion sizes to a separate component
